refactor: Use Go types in demod_get_audio_level wrapper

+++ src/aprs_tt.go
@@ -1638,7 +1638,7 @@ func raw_tt_data_to_app(channel int, msg string) {
 	 */
 
 	if pp != nil {
-		var alevel = demod_get_audio_level(C.int(channel), 0)
+		var alevel = demod_get_audio_level(channel, 0)
 		alevel.mark = -2
 		alevel.space = -2
